requirementsBACKUP.txt
asgiref==3.6.0
aws==0.2.5
awsebcli==3.20.6
bcrypt==4.0.1
boto==2.49.0
botocore==1.29.99
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
crispy-bootstrap4==2022.1
cryptography==41.0.0
decorator==5.1.1
Django==4.1.7
django-crispy-forms==2.0
django-widget-tweaks==1.4.12
dnsdb2==1.1.4
fabric==3.1.0
gunicorn==20.1.0
idna==3.4
# importlib==1.0.4
invoke==2.1.2
jmespath==1.0.1
paramiko==3.2.0
pathspec==0.10.1
postgis==1.0.4
prettytable==3.7.0
psycopg2-binary==2.9.6
pycparser==2.21
PyNaCl==1.5.0
# pypiwin32==223
python-dateutil==2.8.2
# pywin32==306
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
tzdata==2022.7
urllib3==1.26.16
waitress==2.1.2
wcwidth==0.1.9
whitenoise==6.4.0
